# Conflicts:
#	out/production/projektPSiO/code/DostawaStrategia/DostawaKurier.class
#	out/production/projektPSiO/code/DostawaStrategia/DostawaPaczkomat.class
#	out/production/projektPSiO/code/DostawaStrategia/DostawaStrategia.class
#	out/production/projektPSiO/code/Grafika/Koszyk/KoszykFrame.class
#	out/production/projektPSiO/code/Grafika/Koszyk/Test.class
#	out/production/projektPSiO/code/Grafika/Koszyk/Zamowienie.class
#	out/production/projektPSiO/code/InterfejsGraficzny/DodajProduktGUI.class
#	out/production/projektPSiO/code/InterfejsGraficzny/DodajPromocjeGUI$1.class
#	out/production/projektPSiO/code/InterfejsGraficzny/DodajPromocjeGUI.class
#	out/production/projektPSiO/code/InterfejsGraficzny/GUIadmin.class
#	out/production/projektPSiO/code/InterfejsGraficzny/Rejestracja$1.class
#	out/production/projektPSiO/code/InterfejsGraficzny/Rejestracja$2.class
#	out/production/projektPSiO/code/InterfejsGraficzny/Rejestracja.class
#	out/production/projektPSiO/code/InterfejsGraficzny/SklepGUI$1.class
#	out/production/projektPSiO/code/InterfejsGraficzny/SklepGUI$10.class
#	out/production/projektPSiO/code/InterfejsGraficzny/SklepGUI$11.class
#	out/production/projektPSiO/code/InterfejsGraficzny/SklepGUI$12.class
#	out/production/projektPSiO/code/InterfejsGraficzny/SklepGUI$13.class
#	out/production/projektPSiO/code/InterfejsGraficzny/SklepGUI$14.class
#	out/production/projektPSiO/code/InterfejsGraficzny/SklepGUI$2.class
#	out/production/projektPSiO/code/InterfejsGraficzny/SklepGUI$3.class
#	out/production/projektPSiO/code/InterfejsGraficzny/SklepGUI$4.class
#	out/production/projektPSiO/code/InterfejsGraficzny/SklepGUI$5.class
#	out/production/projektPSiO/code/InterfejsGraficzny/SklepGUI$6.class
#	out/production/projektPSiO/code/InterfejsGraficzny/SklepGUI$7.class
#	out/production/projektPSiO/code/InterfejsGraficzny/SklepGUI$8.class
#	out/production/projektPSiO/code/InterfejsGraficzny/SklepGUI$9.class
#	out/production/projektPSiO/code/InterfejsGraficzny/SklepGUI.class
#	out/production/projektPSiO/code/InterfejsGraficzny/UsunProduktGUI.class
#	out/production/projektPSiO/code/Main/Klient.class
#	out/production/projektPSiO/code/Main/Koszyk.class
#	out/production/projektPSiO/code/Main/Main$1.class
#	out/production/projektPSiO/code/Main/Main.class
#	out/production/projektPSiO/code/Main/Sklep.class
#	out/production/projektPSiO/code/Obserwator/Obserwator.class
#	out/production/projektPSiO/code/Obserwator/ObserwatorEmail.class
#	out/production/projektPSiO/code/Obserwator/ObserwatorSMS.class
#	out/production/projektPSiO/code/Obserwator/Podmiot.class
#	out/production/projektPSiO/code/Obserwator/Promocja.class
#	out/production/projektPSiO/code/PlacenieStrategia/PlacenieStrategia.class
#	out/production/projektPSiO/code/PlacenieStrategia/placBlikiem.class
#	out/production/projektPSiO/code/PlacenieStrategia/placKarta.class
#	out/production/projektPSiO/code/Produkt/Bluza.class
#	out/production/projektPSiO/code/Produkt/Koszulka.class
#	out/production/projektPSiO/code/Produkt/Obuwie.class
#	out/production/projektPSiO/code/Produkt/Producent.class
#	out/production/projektPSiO/code/Produkt/Produkt.class
#	out/production/projektPSiO/code/Produkt/Spodnie.class
#	out/production/projektPSiO/code/inputValidate/DaneProduktuException.class
#	out/production/projektPSiO/code/inputValidate/InvalidFileFormatException.class
#	out/production/projektPSiO/code/inputValidate/PlatnoscException.class
#	out/production/projektPSiO/code/inputValidate/PustyKoszykException.class
#	out/production/projektPSiO/code/inputValidate/ZlyAdresException.class

+++ src/code/Produkt/Bluza.java
@@ -9,7 +9,7 @@ public class Bluza extends Produkt implements Serializable {
     private String dekoltBluzy;
     private String krojBluzy;
     
-    public Bluza(String idProduktu, double cena, String nazwa, TreeMap rozmiary, String opis, String material,
+    public Bluza(String idProduktu, float cena, String nazwa, TreeMap rozmiary, String opis, String material,
                  String kolor, Producent producent, boolean czyZKapturem, String dekoltBluzy, String krojBluzy, ImageIcon icon){
         super(idProduktu, cena, nazwa, rozmiary, opis, material, kolor, producent, icon);
         this.czyZKapturem=czyZKapturem;
